+ Update metric_report.sh to be more generic.  Use:
  ./metric_report.sh -e "email1 email2 ..." -p "proj1 proj2"

  Examples:

  ./metric_report.sh -e "jayenne@lanl.gov turners@lanl.gov" \
                     -p "draco clubimc wedgehog"
  ./metric_report.sh -e "jhchang@lanl.gov" -p "draco capsaicin"

  This script must be run from ccscs8 (to allow access to regress's
  /var/tmp directories and regression data).

+ Retire depreciated script 'nightly_cmake_script.sh.'  Use
  master-regression.sh instead.  See xx-crontab files for example
  use.

+++ regression/metrics_report.sh
@@ -1,20 +1,63 @@
 #!/bin/bash
 
 # Collect montly metrics for Jayenne codes and email a report.
+# Use:
+#    <path>/metrics_report.sh -e "email1 [email2]" -p "project1 [project2]"
+
+# Typical use:
+# ./metrics_report.sh -e kellyt@lanl.gov -p "draco clubimc wedgehog"
+# ./metrics_report.sh -e "kgt@lanl.gov jhchang@lanl.gov" -p "draco capsaicin"
+# ./metrics_report.sh -e "jsbrock@lanl.gov barcher@lanl.gov jayenne@lanl.gov" -p "draco clubimc wedgehog"
+
+##---------------------------------------------------------------------------##
+## Support functions
+##---------------------------------------------------------------------------##
+print_use()
+{
+    echo " "
+    echo "Combine code metric data and email report to recipients."
+    echo " "
+    echo "Usage: $0 -e \"email1 [email2]\" -p \"project1 [project2]\""
+    echo " "
+    echo "   [projects] = { draco, clubimc, wedgehog, milagro, capsaicin }."
+    echo "   - Multiple projects or emails must be space delimeted and"
+    echo "     in quotes."
+}
+
+##---------------------------------------------------------------------------##
+## Sanity Checks and Setup
+##---------------------------------------------------------------------------##
+
+if test "${4}x" = "x"; then
+   echo "ERROR: You must provide at least 4 arguments."
+   print_use
+   exit 1
+fi
+
+mach=`uname -n`
+if test "$mach" != "ccscs8.lanl.gov"; then
+   echo "FATAL ERROR: This script must be run from ccscs8."
+   exit 1
+fi
 
 # Ensure the work directory exists
-if ! test -d /var/tmp/kellyt; then
-   mkdir -p /var/tmp/kellyt
+if ! test -d /var/tmp/${USER}; then
+   mkdir -p /var/tmp/${USER}
 fi
 
 # Remove any exising log file.
-logfile=/var/tmp/kellyt/metrics.log
+logfile=/var/tmp/${USER}/metrics.log
 if test -f $logfile; then
    rm $logfile
 fi
+if ! test -d /var/tmp/${USER}; then
+    mkdir /var/tmp/${USER}
+fi
 touch $logfile
 
+##---------------------------------------------------------------------------##
 # Environment setup
+##---------------------------------------------------------------------------##
 
 if test -z "$MODULESHOME"; then
   # This is a new login
@@ -24,138 +67,107 @@ if test -z "$MODULESHOME"; then
 fi
 module load bullseyecoverage
 
+CLOC=/home/regress/draco/regression/cloc
+work_dir=/var/tmp/regress
+
+##---------------------------------------------------------------------------##
+# Process arguments
+##---------------------------------------------------------------------------##
+
+while getopts ":e:p:" opt; do
+case $opt in
+e)
+   recipients=$OPTARG
+   #echo "This report will be emailed to $recipients"
+   ;;
+p)
+   projects=$OPTARG
+   #echo "Processing project(s): $projects"
+   ;;
+:)
+   echo "Option -$OPTARG requires an argument." 
+   print_use
+   exit 1
+   ;;
+\?)
+   print_use
+   exit 1
+   ;;
+esac
+done
+
 # redirect all script output to file
+echo "The metric report will be saved to $logfile "
+echo "and emailed to ${recipients}."
 exec 1>${logfile}
 exec 2>&1
 
+##---------------------------------------------------------------------------##
 # Banner
+##---------------------------------------------------------------------------##
 
-subj=`echo -n "Code Metrics for Draco and Jayenne, "; date`
+# defaults:
+project_name="${projects}"
+pp=""
+# special cases:
+if test "${projects}" = "draco"; then
+   project_name="Draco"
+elif test "${projects}" = "draco clubimc wedgehog"; then
+   project_name="Jayenne"
+   pp=" - Combined report for Draco, ClubIMC and Wedgehog"
+elif test "${projects}" = "draco clubimc wedgehog milagro"; then
+   project_name="Jayenne (with Milagro)"
+   pp=" - Combined report for Draco, ClubIMC, Wedgehog and Milagro"
+elif test "${projects}" = "draco capsaicin"; then
+   project_name="Capsaicin"
+   pp=" - Combined report for Draco and Capsaicin"
+fi
+
+subj=`echo -n "Code Metrics for ${project_name}, "; date`
 
 echo "======================================================================"
-echo ${subj}
+echo "${subj}"
 echo "======================================================================"
-
-# # Draco
-
-# echo " "
-# echo "--------------------"
-# echo "Draco"
-# echo "--------------------"
-# echo " "
-# echo "Lines of code"
-# echo "-------------"
-# cat /home/regress/cmake_draco/Nightly_gcc/Coverage/build/lines-of-code.log
-# echo " "
-# echo "Code coverage"
-# echo "-------------"
-# cd /home/regress/cmake_draco/Nightly_gcc/Coverage/build
-# export COVFILE=`pwd`/CMake.cov
-# export COVDIRCFG=`pwd`/covclass_cmake.cfg
-# cd ../source/src
-# covdir
-
-# # Jayenne - ClubIMC
-
-# echo " "
-# echo " "
-# echo "--------------------"
-# echo "Jayenne - ClubIMC"
-# echo "--------------------"
-# echo " "
-# echo "Lines of code"
-# echo "-------------"
-# cat /home/regress/cmake_jayenne/clubimc/Nightly_gcc/Coverage/build/lines-of-code.log
-# echo " "
-# echo "Code coverage"
-# echo "-------------"
-# cd /home/regress/cmake_jayenne/clubimc/Nightly_gcc/Coverage/build
-# export COVFILE=`pwd`/CMake.cov
-# export COVDIRCFG=`pwd`/covclass_cmake.cfg
-# cd ../source/src
-# covdir
-
-# # Jayenne - Wedgehog
-
-# echo " "
-# echo " "
-# echo "--------------------"
-# echo "Jayenne - Wedgehog"
-# echo "--------------------"
-# echo " "
-# echo "Lines of code"
-# echo "-------------"
-# cat /home/regress/cmake_jayenne/wedgehog/Nightly_gcc/Coverage/build/lines-of-code.log
-# echo " "
-# echo "Code coverage"
-# echo "-------------"
-# cd /home/regress/cmake_jayenne/wedgehog/Nightly_gcc/Coverage/build
-# export COVFILE=`pwd`/CMake.cov
-# export COVDIRCFG=`pwd`/covclass_cmake.cfg
-# cd ../source/src
-# covdir
-
-# Jayenne - Milagro
-
-# echo " "
-# echo " "
-# echo "--------------------"
-# echo "Jayenne - Milagro"
-# echo "--------------------"
-# echo " "
-# echo "Lines of code"
-# echo "-------------"
-# cat /home/regress/cmake_jayenne/milagro/Nightly_gcc/Coverage/build/lines-of-code.log
-# echo " "
-# echo "Code coverage"
-# echo "-------------"
-# cd /home/regress/cmake_jayenne/milagro/Nightly_gcc/Coverage/build
-# export COVFILE=`pwd`/CMake.cov
-# export COVDIRCFG=`pwd`/covclass_cmake.cfg
-# cd ../source/src
-# covdir
-
- # svn diff -r"{2012-01-09}":"{2012-07-09}" --summarize &> ../m_mod_files.log
- #  num_files_changed=`cat ../m_mod_files.log | awk '{print $2}' | sort -u | wc -l`
- #   echo $num_files_changed 
- #  svn log -r"{2012-01-09}":"{2012-07-09}" --verbose &> ../m_msg.log
- #  number_of_commits=`cat ../m_msg.log | grep "Changed paths:" | wc -l`
- #  echo $number_of_commits
- #  files=`find . -name '*.hh' -o -name '*.cc' -o -name '*.txt' -o -name '*.cmake' -o -name '*.in' -o -name '*.h'`
- #  svn annotate $files > ../m_file_list
- #  user_list=`cat ../m_file_list | awk '{print $2}' | sort -u`
- #   echo $user_list 
- #  for name in $user_list; do numlines=`grep $name ../m_file_list | wc -l`; echo "$numlines: $name"; done > ../author_loc
- #   cat ../author_loc | sort -rn
-
-
-# Jayenne - Wedgehog + ClubIMC + Draco
-
 echo " "
 echo " "
 echo "--------------------"
-echo "Jayenne - Combined report for Wedgehog, ClubIMC and Draco"
+echo "${project_name} $pp"
 echo "--------------------"
 echo " "
+# Lines of code report for $projects
+
 echo "Lines of code"
 echo "-------------"
+cmd="${CLOC} --sum-reports "
+for proj in $projects; do
+   cmd="$cmd ${work_dir}/${proj}/Nightly_gcc/Coverage/build/lines-of-code.log "
+done
 # Use grep and head to clean up the output:
-/home/regress/cmake_draco/regression/cloc --sum-reports \
-/home/regress/cmake_draco/Nightly_gcc/Coverage/build/lines-of-code.log \
-/home/regress/cmake_jayenne/clubimc/Nightly_gcc/Coverage/build/lines-of-code.log \
-/home/regress/cmake_jayenne/wedgehog/Nightly_gcc/Coverage/build/lines-of-code.log \
-| grep -v sourceforge | head -n 25
+cmd="$cmd | grep -v sourceforge | head -n 24"
+eval $cmd
+
+# ${CLOC} --sum-reports \
+# ${work_dir}/draco/Nightly_gcc/Coverage/build/lines-of-code.log \
+# ${work_dir}/clubimc/Nightly_gcc/Coverage/build/lines-of-code.log \
+# ${work_dir}/wedgehog/Nightly_gcc/Coverage/build/lines-of-code.log \
+# | grep -v sourceforge | head -n 24
 
 echo " "
 echo "Code coverage"
 echo "-------------"
 echo " "
-export COVFILE=`pwd`/jayenne.cov
-export COVDIRCFG=/home/regress/cmake_draco/regression/covdir.cfg
+export COVFILE=`pwd`/metrics_report.cov
+export COVDIRCFG=/home/regress/draco/regression/covdir.cfg
 if test -f $COVFILE; then
    rm -f $COVFILE
 fi
-covmerge -q --mp --no-banner -c -f $COVFILE /home/regress/cmake_draco/Nightly_gcc/Coverage/build/CMake.cov /home/regress/cmake_jayenne/clubimc/Nightly_gcc/Coverage/build/CMake.cov /home/regress/cmake_jayenne/wedgehog/Nightly_gcc/Coverage/build/CMake.cov
+cmd="covmerge -q --mp --no-banner -c -f $COVFILE "
+for proj in $projects; do
+   cmd="$cmd ${work_dir}/${proj}/Nightly_gcc/Coverage/build/CMake.cov "
+done
+# create the new coverage file via covmerge
+eval $cmd
+# run covdir to generate a report
 covdir
 
 echo " "
@@ -164,9 +176,7 @@ echo "  http://www.bullseye.com/coverage.html#basic_conditionDecision"
 
 # Send the email
 
-/bin/mailx -s "${subj}" kellyt@lanl.gov < ${logfile}
+echo /bin/mailx -s "${subj}" ${recipients} < ${logfile}
+/bin/mailx -s "${subj}" ${recipients} < ${logfile}
+#/bin/mailx -s "${subj}" kellyt@lanl.gov < ${logfile}
 #/bin/mailx -s "${subj}" jsbrock@lanl.gov barcher@lanl.gov jayenne@lanl.gov < ${logfile}
-
-# jayenne@lanl.gov
-# jsbrock@lanl.gov
-# barcher@lanl.gov
